requirements.txt
streamlit
python-dotenv
google-generativeai
google-adk
IPython
